use URLSearchParams when grabbing events.
