+++ types.ts
@@ -18,3 +18,21 @@ export interface SearchCriteria {
   page: number;
   collectionSlug: string;
 }
+
+export interface NFTEvent {
+  assetName: string;
+  assetDescription: string;
+  assetImgUrl: string;
+  assetUrl: string;
+  collectionName: string;
+  collectionImgUrl: string;
+  collectionUrl: string;
+  collectionDescription: string;
+  date: string;
+  from: string;
+  to: string;
+  action: Action;
+  key: string;
+  transactionHash: string; // event's from opensea can have the same transaction hash, can't just store this in key (ex. mintpass transaction)
+  price?: number;
+}
